+++ CHANGELOG.md
@@ -11,3 +11,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Component blocks setup.
 - Initial component with postalCode calculator
+- Shipping result with Shipping Options